Switch the test install script to be version specific
WordPress and the test lib are installed to version
specific dirs and symlinked in to allow multiple
versions to be installed (effectively cached) locally.

The script now installs WordPress and the WordPress
test lib to a version specific directory, in order that
we don't run into issues where classes or functions
are not defined in WordPress but are used in the
test lib.