Don't error if user has already been deleted
We just want the remove-user command to ensure the support user in question has been removed. If they don't exist, we can note that in the output, but shouldn't error because it causes vipd host actions to fail. An easy way to get into this scenario is just to add two support users in a row. Both times we'll queue a second action to delete the user later, but the second `remove-user` will fail.