Add VIP Support Email Aliases (#73)
* Add support for VIP Support email aliases

Masks `vip-support+<username>@automattic.com` addresses to appear as
`vip-support@automattic.com` instead, primarily for the Users lists
and for Gravatar.

Added to help clearly brand VIP staff who don't use a8c email addresses.

* Add unit tests for verifying VIP Support email aliases

* split is_vip_support_email_alias tests into valid & invalid

makes it more clear when a tests fails wether it was the valid emails or
invalid emails.

Cleaned up structure by using dataProvider.

* Add msg param to support email constant test

On fail, this test does not specify the issue, only that it failed.
It lacked context and am actionable message.

Clear description of the problem and instructions are now provided on
fail.

* support email alias pattern moved to constant

breaking apart email to create a dynamic regex pattern was convoluted
and not very readable.

moved regex pattern to a constant

* Add unit test for verifying VIP support email alias pattern

If the constant ever changes, in/valid email examples need to change as
well. This test attempts to make that clear.

* change part of VIP Support email regex pattern

`.*` is a bit too loose and may cause complications. switching to something a little less vague

Co-Authored-By: davidsword <david@davidsword.ca>

* Update test for VIP Support email regex pattern

slight alteration to regex pattern to match
adjustments made to pattern in class-vip-support-user.php

* Change VIP Support emails Gravtar to be constant instead of dynamic

Running md5() on the email is unessisary as the email and Gravatar
are unlikly to ever change

* Remove unnecessary md5

No longer needed since we hard-code the Gravatar URL.

* Fix dockblock

props @GaryJones